5401: fixes after merge with release/4
